updated: cleaning up the unit test and added the possibility to compile multiple tags paths via compile
