+++ lib/browser/compiler/compile.js
@@ -2,7 +2,7 @@ import observable from 'riot-observable'
 import compiler from 'riot-compiler'
 import * as r from '../index'
 import { T_STRING } from '../global-variables'
-import { extend, isFunction, $$, getAttr, mkEl, isObject } from '../util'
+import { extend, isFunction, isArray, $$, getAttr, mkEl, isObject } from '../util'
 
 
 export function mount(a, b, c) {
@@ -110,8 +110,18 @@ export var compile = (function () {
         if (fn) fn(js, str, opts)
       })
 
-    }
-    else {
+    } else if (isArray(arg)) {
+      var i = arg.length
+      // `riot.compile([urlsList] [, callback][, options])`
+      arg.forEach(function(str) {
+        GET(str, function (str, opts, url) {
+          var js = compiler.compile(str, opts, url)
+          globalEval(js, url)
+          i --
+          if (!i && fn) fn(js, str, opts)
+        })
+      })
+    } else {
 
       // `riot.compile([callback][, options])`
       if (isFunction(arg)) {
