+++ package.json
@@ -76,7 +76,7 @@
     "rollup-plugin-babel": "^2.4.0",
     "rollup-plugin-commonjs": "^2.2.1",
     "rollup-plugin-npm": "^1.4.0",
-    "rollup-plugin-riot": "^0.3.0",
+    "rollup-plugin-riot": "^1.0.0-alpha.1",
     "sinon": "^1.17.4",
     "sinon-chai": "^2.8.0",
     "uglify-js": "latest"
