+++ test/specs/browser/_tags-bootstrap.js
@@ -1,40 +1,14 @@
 
 // bootstrap all the tags
 loadTagsAndScripts([
-  'tag/~custom-parsers.tag',
-  'tag/loop.tag',
-  'tag/loop-child.tag',
-  'tag/loop-combo.tag',
-  'tag/loop-reorder.tag',
-  'tag/loop-manip.tag',
-  'tag/loop-object.tag',
-  'tag/loop-tag-instances.tag',
-  'tag/loop-numbers-nested.tag',
-  'tag/loop-nested-strings-array.tag',
-  'tag/loop-events.tag',
-  'tag/loop-sync-options-nested.tag',
-  'tag/loop-inherit.tag',
-  'tag/loop-root.tag',
-  'tag/loop-double-curly-brackets.tag',
-  'tag/loop-conditional.tag',
+
   'tag/table-multibody.tag',
-  'tag/loop-cols.tag',
   'tag/nested-child.tag',
-  'tag/loop-option.tag',
-  'tag/loop-optgroup.tag',
-  'tag/loop-optgroup2.tag',
-  'tag/loop-position.tag',
-  'tag/loop-arraylike.tag',
-  'tag/loop-ids.tag',
-  'tag/loop-unshift.tag',
-  'tag/loop-virtual.tag',
-  'tag/loop-null-items.tag',
+
   'tag/table-data.tag',
-  'tag/table-loop-extra-row.tag',
   'tag/table-thead-tfoot.tag',
   'tag/table-thead-tfoot-nested.tag',
-  'tag/loop-named.tag',
-  'tag/loop-single-tags.tag',
+
   'tag/timetable.tag',
   'tag/raw-contents.tag',
   'tag/virtual-nested-unmount.tag',
@@ -66,8 +40,7 @@ loadTagsAndScripts([
   // recursive tags
   'tag/treeview.tag',
 
-  // sync the loop options
-  'tag/loop-sync-options.tag',
+
 
   // check if the events get triggered correctly
   'tag/events.tag',
@@ -86,24 +59,14 @@ loadTagsAndScripts([
 
   'tag/reserved-names.tag',
 
-  'tag/obj-key-loop.tag',
-
   'tag/should-update.tag',
 
-  'tag/virtual-no-loop.tag',
-  'tag/virtual-yield-loop.tag',
+
   'tag/dynamic-data-is.tag',
 
-  'tag/loop-protect-internal-attrs.tag',
-
-  'tag/loop-svg-nodes.tag',
 
   // these tags will be not autoinjected in the DOM
   // that's what `name = false` means
-  {
-    path: 'tag/ploop-tag.tag',
-    name: false
-  },
   {
     path: 'tag/timer.tag',
     name: false
@@ -120,10 +83,7 @@ loadTagsAndScripts([
     path: 'tag/inner-html.tag',
     name: false
   },
-  {
-    path: 'tag/outer.tag',
-    name: false
-  },
+
   {
     path: 'tag/named-child.tag',
     name: false
