+++ test/tag/style-tag5.tag
@@ -1,0 +1,10 @@
+<style-tag5>
+  <style scoped="scoped">
+    p {border: solid 3px black;}
+  </style>
+  <style>
+    p {border: solid 1px black}
+    #style4 {border: solid 2px black;}
+  </style>
+  <p></p>
+</style-tag5>
